backport r150268.
svn path=/branches/mono-2-6/mcs/; revision=150269

+++ mcs/class/System.Json/System.Json/JsonReader.cs
@@ -155,7 +155,7 @@ namespace System.Json
 			}
 
 			int c;
-			int val = 0;
+			decimal val = 0;
 			int x = 0;
 			bool zeroStart = PeekChar () == '0';
 			for (; ; x++) {
